GoogleAuth in hooks + baseUrl: src
