+++ package.json
@@ -52,6 +52,7 @@
     "prettier": "^2.8.0",
     "tailwindcss": "^3.2.4",
     "typescript": "^4.9.3",
-    "vite": "^3.2.3"
+    "vite": "^3.2.3",
+    "vite-tsconfig-paths": "^4.0.5"
   }
 }
