+++ tsconfig.json
@@ -15,7 +15,7 @@
     "isolatedModules": true,
     "noEmit": true,
     "jsx": "react-jsx",
-    "baseUrl": "src"
+    "baseUrl": "src",
   },
   "include": ["src", "index.d.ts", "/*.ts", "/*.tsx"],
   "exclude": ["node_modules", "/*.js", "/*.jsx"],
